MA-10983 Pad HDMI firmware after iMX8QM bootloader
As iMX8QM need to load HDMI firmware in bootloader
to make HDMI interface available for both u-boot's splash
and Linux. So pad it after u-boot.bin will be the best solution
to make u-boot fetch the firmware data.

This patch modify soc.mak and judge when there is HDMI firmware
and pad it after u-boot.bin. This will enlarge u-boot by 128KB.

Signed-off-by: Haoran.Wang <elven.wang@nxp.com>