Add support for QXB0
This adds the option rev in order to differentiate between QX A0 and B0. The
container header has changed a lot and therefore a new implementation of the
imx8qx file had to be added. The option append is also added to allow the
passing of the first container as a blob file. The signature block for the
container created, needs to be updated when the signature block is appended by
CST. The hash is set by default to sha386 and the IV is not handled at all at
this point.

Signed-off-by: Abel Vesa <abel.vesa@nxp.com>