imx8qxb0: Add data option to be used by mfgtool
This allows passing binary blobs to be appended to the output container file.
There can be as many data blobs appended as needed. The format of this option
is as follows:

 -data <file> <load_addr>

Also added a new target to iMX8QX/soc.mak to be used as an example of
how to use.

Signed-off-by: Abel Vesa <abel.vesa@nxp.com>
Reviewed-by: Ye Li <ye.li@nxp.com>
Reviewed-by: Frank Li <frank.li@nxp.com>