imx8m: Add iMX8M Nano support
Since the IVT offset is changed on this SoC. Added a -version parameter
to pass the v1 or v2 to mkimage_imx8.
v1 is for iMX8MQ and iMX8MM
v2 is for iMX8M Nano (iMX8MN)

Also the ROM on iMX8MN does not support DCD nor plugin. Checks the
parameters if they are used.

Additional, on Nano the addresses of SPL and ATF are changed to OCRAM.

Signed-off-by: Ye Li <ye.li@nxp.com>