Revert "imx8: Change default M4 partition to 1"
This reverts commit 5e9a4be871e0e53b9ec6700a9f793ce163eb9c12.

Clarify the logic for -p parameter as below. So consider the case that
M4 and AP share in one partition, secondary CPU need set -p0, this secondary
CPU could be AP either M4.
It is better to revert the patch to avoid break this case.
Our single M4 targets have added -p1, so no impact to them.

"
 - Without the boot flag there is only a partition 1. One CPU must be p1.
   Others must be p0 (indicates secondary CPU).  The default was no boot flag,
   AP is p1, M4s p0. If no AP then one of the M4 needs to be p1 (but not both).
   With the bootflag there is 1, 3, and 4 (on QM). AP should be p1 (the default)
   and M4_0 is p3 and M4_1 is p4.
"

Signed-off-by: Ye Li <ye.li@nxp.com>