MLK-22571-6 imx8m: Add FIT DEK Blob as an image entry in FIT image
The CST generates a random DEK when encrypting images which must be
encapsulated into a CAAM BLOB.

During the decryption processes the HABv4 code decapsulate the blob which
must be loaded in a specific memory address.

This patch is adding a dummy dek blob as an image entry in FIT image, the
address is hardcoded to DEK_BLOB_LOAD_ADDR=0x40400000 and can be modified
according to user specific use case.

Details about HABv4 encrypted boot and how to replace dummy DEK Blob can be
found in U-Boot documentation at doc/imx/habv4/

Reviewed-by: Ye Li <ye.li@nxp.com>
Signed-off-by: Breno Lima <breno.lima@nxp.com>