MA-15300-2 boot with A72 to save boot time in Android Automotive
change the parameter of "a53" to be "a72" when invoke mkimage_imx8

the rule for target "flash_b0_spl_container_m4_0_1_trusty" is a copy of
"flash_linux_m4", the target is renamed, and "a53" in the recipe is
modified to "a72".

leave the target for "mx8qm-ahab-container.img" to use "a53", spl won't
touch this parameter when load and execute the code in this file.

Change-Id: If302a8d6371193783890d6e198e56ba069737508
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>