MLK-25292 iMX8M: Remove the usage of IVT reserved1 field
On iMX8MQ B2, ROM will check IVT header and requires the reserved fields
in IVT to be 0.
However, in imx-mkimage we set the reserved1 field for the offset from
second boot image to SPL boot image, which was used for mfgtool but
has deprecated. So remove it to fix the boot failure on iMX8MQ B2.

Signed-off-by: Ye Li <ye.li@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>