iMX8QM/QX/DXL/ULP: Add core id to -data option
Some use cases need data images to be loaded inside the M4 TCML.
This requires the core id of the image header to be set to cm4
(ROM code will refuse to load the TCM if the core id is not the CM4).
So allow the core id to be passed to the -data option to handle this
case.

Signed-off-by: Seb Haezebrouck <sebastien.haezebrouck@nxp.com>
Signed-off-by: Ye Li <ye.li@nxp.com>