Use common boottools location [YOCIMX-7189]
The Yocto daily build currently builds one set of boottools and stores
it in a common location in Nexus. Then, for each of the three image
builds, Linux_IMX_Core, Linux_IMX_Regression_Next_Kernel, and
Linux_IMX_Full, the boottools are downloaded from Nexus and uploaded
to the image location in Nexus.

Obviously, the image location copies are extremely costly. In order to
drop the copies, adapt the makefiles to use the common location, which
is stored in each image location in the file boottools_location.txt.

The new logic is applied only for builds done after this change. This
is facilitated by adding a version to the build, stored in the file
boottools_build_version.txt.

Also, for maximum flexibility, a new command line option is added to
disable the logic `USE_COMMON_LOCATION=false`.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>