Update Yocto build name to Common [YOCIMX-7230]
The Yocto build called `BootTools` does in fact contain other
components. Rename it to Common.

The change does maintain compatibility for the case that the new script
is used with an old build. In that case, the new file names will not be
found, and the old file names are used instead.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>