imx95: Clean up targets in soc.mak and update Readme.imx95
Add a dedicated SOC type for iMX95 to avoid conflict with iMX93 setting,
and remove flash_xxx_oei targets which are replaced with "OEI=YES".

So the build commands are simplified as below, for example:

To use SPL to intialize DDR:
<singleboot>	make SOC=iMX95 flash_singleboot
<lpboot>	make SOC=iMX95 flash_lpboot_sm_a55

To use OEI to initialize DDR:
<singleboot>	make SOC=iMX95 OEI=YES flash_singleboot
<lpboot>	make SOC=iMX95 OEI=YES flash_lpboot_sm_a55

Refer the Readme.imx95 for details

Note: ATF and u-boot address are updated to align with SCMI version.
Non-SCMI version needs to adjust to align with SCMI.
ATF_LOAD_ADDR is 0x204A0000
UBOOT_LOAD_ADDR is 0x90200000

Signed-off-by: Ye Li <ye.li@nxp.com>
Signed-off-by: Viorel Suman <viorel.suman@nxp.com>