iMX95: add suport to append DDR PHY training data data to DDR OEI
For DDR quick boot flow the training data needs to be carried with
DDR OEI - add support for this. By default the usual training
DDR OEI is built. QB DDR OEI is built if LPDDR_FUNC=qboot is
defined before running make command. The training data is
expected to be present in "iMX95" folder as "qb_data.bin" file.

Signed-off-by: Viorel Suman <viorel.suman@nxp.com>